% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{area_ordered}
\alias{area_ordered}
\title{Area under curve order relation among univariate functional data}
\usage{
area_ordered(fData, gData)
}
\arguments{
\item{fData}{the first univariate functional dataset containing elements to
be compared}

\item{gData}{the second univariate functional dataset containing elements to
be compared}
}
\description{
It implements the area under curve order relation among univariate f
 unctional data, that is the pre-order relation obtained by comparing the
 area under curve maxima of two different functional data.
}
\details{
It accepts two fData objects, that can have either same sample size (number
 of elemnets), or just one element. In the former case the comparison is made
 element-wise between corresponding observations of the two datasets; in the
 latter case, the dataset with just one observation is used cyclically
 against the other one (if need be).
}

