% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BandDepths.R
\name{BD}
\alias{BD}
\title{Band Depth for univariate functional data}
\usage{
BD(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{BD} computes the Band Depth of a functional dataset
}

