% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BandDepths.R
\name{BD_relative}
\alias{BD_relative}
\title{Relative Band Depth of functions in a functional dataset
\code{BD_relative} computes the Band Depth (BD) of a functional dataset (Data_target) with respect to another (Data_reference)}
\usage{
BD_relative(Data_target, Data_reference)
}
\arguments{
\item{Data_target}{is the dataset containing signals whose depths are to be computed with respect to the reference population.}

\item{Data_reference}{is the dataset containing the reference population used to measure the centrality through depths. In particular, the depths of these functions will not be computed.}
}
\description{
Relative Band Depth of functions in a functional dataset
\code{BD_relative} computes the Band Depth (BD) of a functional dataset (Data_target) with respect to another (Data_reference)
}

