% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_kendall}
\alias{cor_kendall}
\title{Kendall's tau correlation coefficient for a bivariate functional dataset}
\usage{
cor_kendall(mfD, ordering = "max")
}
\arguments{
\item{mfD}{the bivariate functional dataset whose Kendall's tau
coefficient must be computed}

\item{ordering}{the ordering relation to use on functional observations,
either "max" for the maximum relation or "area" for the area under the curve
relation}
}
\description{
It implements the computation of the Kendall's tau correlation coefficient
 for bivariate functional data, with the prescribed order relation (either
 max or area-under-curve)
}

