% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_spearman}
\alias{cor_spearman}
\title{Spearman's correlation coefficient for a bivariate functional dataset}
\usage{
cor_spearman(mfD, ordering = "MEI", ...)
}
\arguments{
\item{mfD}{the bivariate functional dataset whose Kendall's tau
coefficient must be computed}

\item{ordering}{the ordering relation to use on functional observations,
either "MEI" or "MHI"}

\item{...}{additional parameters to be passed to rank function}
}
\description{
It implements the computation of the Spearmans's correlation coefficient
 for bivariate functional data, with the prescribed order relation (either
 MEI or MHI)
}

