% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_spearman_accuracy}
\alias{cor_spearman_accuracy}
\title{Bootstrap Spearman's correlation coefficient for multivariate functional data}
\usage{
cor_spearman_accuracy(mfD, ordering = "MEI", bootstrap_iterations = 1000,
  verbose = FALSE)
}
\description{
FINISH DOC
}
