% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthGramPlot.R
\name{depthGramPlot}
\alias{depthGramPlot}
\title{Plot DepthGram}
\usage{
depthGramPlot(
  DG,
  limits = F,
  ids = NULL,
  print = F,
  plotly = F,
  plot.title = "",
  shorten = T,
  col = NULL,
  pch = 19,
  sp = 2,
  st = 4,
  sa = 10,
  text.labels = ""
)
}
\arguments{
\item{DG}{output of the depthGram function.}

\item{limits}{if TRUE the empirical limits for outlier detection are drawn. Default is FALSE.}

\item{ids}{labels for data points.}

\item{print}{if TRUE the graphical output is optimized for printed version. Default is FALSE.}

\item{plotly}{if TRUE the graphical output is displayed as an interactive plotly object. Default is FALSE.}

\item{plot.title}{character with the main title for plot.}

\item{shorten}{If labels must be shorten to 15 characters. Default is TRUE.}

\item{col}{colors for the plot. Default is NULL.}

\item{pch}{point shape.}

\item{sp}{point size.}

\item{st}{label size.}

\item{sa}{axis title sizes.}

\item{text.labels}{labels for the individuals. text.labels is overridden if limits=T, for which only outliers labels are shown.}
}
\value{
A list with the following items:
\itemize{
\item p - list with all the depthGram plots.
\item out - outliers detected.
\item colors - used colors for plotting.
}
}
\description{
Function to plot the 3 DepthGram representations from the output of the depthGram function
}
\examples{

N = 2e2
P = 1e3
grid = seq( 0, 1, length.out = P )
Cov = exp_cov_function( grid, alpha = 0.3, beta = 0.4 )

Data <- list()
Data[[1]] <- generate_gauss_fdata( N, centerline = sin( 2 * pi * grid ), Cov = Cov )
Data[[2]] <- generate_gauss_fdata( N, centerline = sin( 2 * pi * grid ), Cov = Cov )
names <- paste0("id_", 1:nrow(Data[[1]]))
DG <- depthGram(Data, marg.out=TRUE, ids=names)
depthGramPlot(DG)
}
\references{
Aleman-Gomez, Y., Arribas-Gil, A., Desco, M. Elias-Fernandez, A., and Romo, J. (2021). "Depthgram: Visualizing Outliers in High Dimensional Functional Data with application to Task fMRI data exploration". \url{arXiv:2103.08874}
}
