% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indexes.R
\name{EI}
\alias{EI}
\title{Epigraph Index of functions in a functional dataset}
\usage{
EI(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{EI} computes the Epigraph Index of a functional dataset
}

