% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbplot.R
\name{fbplot.fData}
\alias{fbplot.fData}
\title{Functional boxplot for univariate functional data}
\usage{
\method{fbplot}{fData}(Data, Depths = "MBD", adjust = FALSE,
  display = TRUE, ..., verbose = FALSE)
}
\arguments{
\item{Data}{the univariate functional dataset whose functional boxplot is
desired}

\item{Depths}{either a vector containing the depths for each element of the
dataset, or a string containing the name of the method you want to use to
compute it. In this case the name of the method must be included in the
caller's environment}

\item{adjust}{either FALSE if you would like the default value for the
inflation factor, F = 1.5, to be used, or a list specifying the parameters
required by the adjustment.}

\item{display}{either a logical value indicating wether you want the
functional boxplot to be displayed, or the number of the graphical devices
where you want the boxplot to be plotted.}

\item{...}{additional graphical parameters to be used in plotting functions}

\item{verbose}{logical flag to enable verbosity in the adjustment process}
}
\description{
\code{fbplot.fData} displays the functional boxplot of a dataset of functional data.
}

