% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbplot.R
\name{fbplot.mfData}
\alias{fbplot.mfData}
\title{Functional boxplot for multivariate functional data}
\usage{
\method{fbplot}{mfData}(Data, Depths = list(def = "MBD", weights = "uniform"),
  Fvalue = 1.5, adjust = FALSE, display = TRUE, xlab = NULL,
  ylab = NULL, main = NULL, ..., verbose = FALSE)
}
\arguments{
\item{Data}{the univariate functional dataset whose functional boxplot is
desired}

\item{Depths}{either a vector containing the depths for each element of the
dataset, or a string containing the name of the method you want to use to
compute it. In this case the name of the method must be included in the
caller's environment}

\item{Fvalue}{the value of the inflation factor F, default is F = 1.5,}

\item{adjust}{either FALSE if you would like the default value for the
inflation factor, F = 1.5, to be used, or a list specifying the parameters
required by the adjustment.}

\item{display}{either a logical value indicating wether you want the
functional boxplot to be displayed, or the number of the graphical devices
where you want the boxplot to be plotted.}

\item{xlab}{the label to use on the x axis when displaying the functional
boxplot}

\item{ylab}{the label to use on the y axis when displaying the functional
boxplot}

\item{main}{the main title to used when displaying the functional boxplot}

\item{...}{additional graphical parameters to be used in plotting functions}

\item{verbose}{logical flag to enable verbosity in the adjustment process}
}
\description{
\code{fbplot.fData} displays the functional boxplot of a dataset of
functional data.
}

