% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{fDColorPalette}
\alias{fDColorPalette}
\title{A set of fancy color to plot fData objects}
\usage{
fDColorPalette(N, range = c(0, 360), alpha = 0.8, ...)
}
\arguments{
\item{N}{number of different colors (ideally, functional observations)}

\item{range}{the range of hues to be used}

\item{alpha}{the alpha channel parameter of the colors (transparency)}

\item{...}{additional parameters to be passed to \code{scales::hue_pal}}
}
\description{
A set of fancy color to plot fData objects
}

