% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indexes.R
\name{HI}
\alias{HI}
\title{Hypograph Index of functions in a functional dataset}
\usage{
HI(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{EI} computes the Modified Index of a functional dataset
}

