% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indexes.R
\name{HRD}
\alias{HRD}
\title{Half-region depth}
\usage{
HRD(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{HRD} computes the Half Region Depth of a functional dataset
}

