% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{max_fData}
\alias{max_fData}
\title{\code{max_fData} maximum of an univariate functional dataset}
\usage{
max_fData(fData, ..., which = FALSE)
}
\arguments{
\item{fData}{the functional dataset containing elements whose maxima have to
be computed}

\item{...}{additional parameters to pass to max}

\item{which}{logical flag specifying whether the grid values where maxima are
fulfilled have to be returned too}
}
\description{
It computes the maximum value of each element of the functional dataset,
 and optionally returns also the value of the grid where they are fulfilled
}

