% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BandDepths.R
\name{MBD}
\alias{MBD}
\title{Modified Band Depth for univariate funcitonal data}
\usage{
MBD(Data, manage_ties = FALSE)
}
\arguments{
\item{Data}{is the dataset of functional observations whose MBD are to be computed. Its rows
should be the functional observations, while the columns must be time-point evaluations of such
observations}

\item{manage_ties}{is a flag specifying if the method should accomodate for the presence of ties
in the dataset.}
}
\description{
\code{MBD} Computes the Modified Band Depth (MBD) of elements of a functional dataset
}

