% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BandDepths.R
\name{MBD_relative}
\alias{MBD_relative}
\title{Relative Modified Band Depth of functions in a functional dataset
\code{MBD_relative} computes the Modified Band Depth (MBD) of a functional dataset (Data_target) with respect to another (Data_reference)}
\usage{
MBD_relative(Data_target, Data_reference)
}
\arguments{
\item{Data_target}{is the dataset containing signals whose depths are to be computed with respect to the reference population.}

\item{Data_reference}{is the dataset containing the reference population used to measure the centrality through depths. In particular, the depths of these functions will not be computed.}
}
\description{
Relative Modified Band Depth of functions in a functional dataset
\code{MBD_relative} computes the Modified Band Depth (MBD) of a functional dataset (Data_target) with respect to another (Data_reference)
}

