% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{mean_fData}
\alias{mean_fData}
\title{\code{mean_fData} method to compute the sample mean of a fData object.}
\usage{
mean_fData(fData)
}
\arguments{
\item{fData}{the functional data object representing the dataset}
}
\description{
It computes the \bold{cross-sectional} mean of a univariate functional
dataset, i.e., its time-by-time sample mean.
}

