% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{median.fData}
\alias{median.fData}
\title{\code{median.fData} method to compute the sample median of a fData object.}
\usage{
\method{median}{fData}(fData, type = "MBD")
}
\arguments{
\item{fData}{the functional data object containing the dataset}

\item{type}{the type of depth definition you want to use to define the sample median (default is MBD)}
}
\description{
\code{median.fData} method to compute the sample median of a fData object.
}

