% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{median_fData}
\alias{median_fData}
\title{\code{median_fData} method to compute the sample median of a fData object.}
\usage{
median_fData(fData, type = "MBD")
}
\arguments{
\item{fData}{the functional data object containing the dataset}

\item{type}{depth definition to use in order to find the sample median
(default is MBD)}
}
\description{
It computes the depth-based median of a univariate functional dataset, i.e.
it finds the deepest element of the functional dataset according to a
specified definition of depth.
}

