% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indexes.R
\name{MEI}
\alias{MEI}
\title{Modified Epigraph Index of functions in a functional dataset}
\usage{
MEI(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{MEI} computes the Modified Epigraph Index of a functional dataset
}

