% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfD_healthy.R
\docType{data}
\name{mfD_healthy}
\alias{mfD_healthy}
\title{ECG trace of healthy subjects}
\format{An object of class \code{mfData} of length 6.}
\usage{
mfD_healthy
}
\description{
A dataset containing the 8-Lead ECG traces of 150 healthy subjects. They can be used to compare the
signals of pathological subjects stored in \code{mfD_LBBB} and \code{mfD_RBBB} objects.
}
\details{
The 8 leads are, in order, V1, V2, V3, V4, V5, D1 and D2. The signals have been registered and
smoothed over an evenly spaced grid of 1024 time points at 1kHz.
}
\keyword{datasets}
