% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfD_LBBB.R
\docType{data}
\name{mfD_LBBB}
\alias{mfD_LBBB}
\title{ECG trace of subjects suffering from Left-Bundle-Branch-Block (LBBB)}
\format{A \code{\link{mfData}} object.}
\usage{
mfD_LBBB
}
\description{
A dataset containing the 8-Lead ECG traces of 50 subjects suffering from
Left-Bundle-Branch-Block (LBBB), a cardiac pathology affecting the conduction process and resulting
in some peculiar distortions of the ECG.
}
\details{
The 8 leads are, in order, V1, V2, V3, V4, V5, D1 and D2. The signals have been registered and
smoothed over an evenly spaced grid of 1024 time points at 1kHz.
}
\keyword{datasets}
