% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfD_RBBB.R
\docType{data}
\name{mfD_RBBB}
\alias{mfD_RBBB}
\title{ECG trace of subjects suffering from Right-Bundle-Branch-Block (RBBB)}
\format{An object of class \code{mfData} of length 6.}
\usage{
mfD_RBBB
}
\description{
A dataset containing the 8-Lead ECG traces of 50 subjects suffering from
Right-Bundle-Branch-Block (RBBB), a cardiac pathology affecting the conduction process and resulting
in some peculiar distortions of the ECG.
}
\details{
The 8 leads are, in order, V1, V2, V3, V4, V5, D1 and D2. The signals have been registered and
smoothed over an evenly spaced grid of 1024 time points at 1kHz.
}
\keyword{datasets}
