% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{mfData}
\alias{mfData}
\title{\code{mfData} generic class for (multivariate) functional data.}
\usage{
mfData(grid, Data_list)
}
\arguments{
\item{grid}{the (evenly spaced) grid over which the functional data is defined}

\item{Data_list}{a list containing the time-by-time values of functional
dataset, where each node of the list (possibly named) represents a particular
dimension of the multivariate dataset.}
}
\description{
\code{mfData} generic class for (multivariate) functional data.
}

