% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indexes.R
\name{MHI}
\alias{MHI}
\title{Modified Hypograph Index of functions in a functional dataset}
\usage{
MHI(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{MEI} computes the Modified Hypograph Index of a functional dataset
}

