% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indexes.R
\name{MHRD}
\alias{MHRD}
\title{Modified Half-region depth}
\usage{
MHRD(Data)
}
\arguments{
\item{Data}{a matrix-like dataset of functional data, with observations as rows and time points as columns}
}
\description{
\code{MHRD} computes the Modified Half Region Depth of a functional dataset
}

