% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{min_fData}
\alias{min_fData}
\title{\code{min_fData} minimum of an univariate functional dataset}
\usage{
min_fData(fData, ..., which = FALSE)
}
\arguments{
\item{fData}{the functional dataset containing elements whose minima have to
be computed}

\item{...}{additional parameters to pass to max}

\item{which}{logical flag specifying whether the grid values where minima are
fulfilled have to be returned too}
}
\description{
It computes the minimum value of each element of the functional dataset,
 and optionally returns also the value of the grid where they are fulfilled
}

