% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multivariate_depths.R
\name{multiBD}
\alias{multiBD}
\title{Band Depth for multivariate functional data}
\usage{
multiBD(listData, weights = "uniform")
}
\arguments{
\item{listData}{specifies the multivariate functional dataset. It is a list
of 2-dimensional matirices containing as rows the observations and as columns
the measurements of the functional data over the continuous variable (e.g.
time)}

\item{weights}{either a set of weights (of the same length of \code{listData}
) or the string \code{uniform} specifying that a set of uniform weights
(1 / L, where L is the number of dimensions of the functional dataset and
thus the length of \code{listData}) is to be used.}
}
\description{
\code{BD} Computes the Band Depth (MBD) of elements of a multivariate
functional dataset
}

