% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multivariate_depths.R
\name{multiMBD}
\alias{multiMBD}
\title{Modified Band Depth for multivariate functional data}
\usage{
multiMBD(listData, weights = "uniform", manage_ties = FALSE)
}
\arguments{
\item{listData}{specifies the multivariate functional dataset. It is a list
of 2-dimensional matirices containing as rows the observations and as columns
the measurements of the functional data over the continuous variable (e.g.
time)}

\item{weights}{either a set of weights (of the same length of \code{listData}
) or the string \code{uniform} specifying that a set of uniform weights
(1 / L, where L is the number of dimensions of the functional dataset and
thus the length of \code{listData}) is to be used.}

\item{manage_ties}{a logical flag specifying whether the check for ties and
the relative treatment is to be carried out while computing the MBDs in each
dimension. It is directly passed to \code{MBD}.}
}
\description{
\code{MBD} Computes the Band Depth (MBD) of elements of a multivariate
functional dataset
}

