% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot.fData}
\alias{plot.fData}
\title{\code{plot.fData} specialised method to plot univariate functional data..}
\usage{
\method{plot}{fData}(fData, lty = 1, col = NULL, ...)
}
\arguments{
\item{fData}{the univariate functional data object}

\item{lty}{lty graphical parameter to be used in plotting functions}

\item{col}{colors to be used in plotting functions}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
\code{plot.fData} specialised method to plot univariate functional data..
}

