% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot.mfData}
\alias{plot.mfData}
\title{Specialised method to plot multivariate functional data.}
\usage{
\method{plot}{mfData}(x, ...)
}
\arguments{
\item{x}{the multivariate functional data object}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
Specialised method to plot multivariate functional data.
}

