% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot_fData}
\alias{plot_fData}
\title{Specialised method to plot univariate functional data..}
\usage{
plot_fData(fData, col = NULL, lty = 1, ...)
}
\arguments{
\item{fData}{the univariate functional data object}

\item{col}{colors to be used in plotting functions}

\item{lty}{lty graphical parameter to be used in plotting functions}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
Specialised method to plot univariate functional data..
}

