% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot_mfData}
\alias{plot_mfData}
\title{Specialised method to plot multivariate functional data.}
\usage{
plot_mfData(mfData, ylab = NULL, main = NULL, col = NULL, lty = 1, ...)
}
\arguments{
\item{mfData}{the multivariate functional data object}

\item{ylab}{either a string or a list of stirngs to be used as y labels for
each window in the plot panel, default is null}

\item{main}{either a string or a list of stirngs to be used as main title for
each window in the plot panel, default is null}

\item{col}{colors to be used in plotting functions}

\item{lty}{lty graphical parameter to be used in plotting functions}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
Specialised method to plot multivariate functional data.
}

