% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot_mfData_default}
\alias{plot_mfData_default}
\title{Default method to plot multivariate functional data.}
\usage{
plot_mfData_default(x, type = "l", lty = 1, col = fDColorPalette(x$N),
  xlab = NULL, ylab = NULL, main = NULL, ...)
}
\arguments{
\item{x}{the multivariate functional data object}

\item{lty}{lty graphical parameter to be used in plotting functions}

\item{col}{colors to be used in plotting functions}

\item{xlab}{the x label to add to each plot window}

\item{ylab}{either a string or a list of stirngs to be used as y labels for
each window in the plot panel, default is null}

\item{main}{either a string or a list of stirngs to be used as main title for
each window in the plot panel, default is null}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
Default method to plot multivariate functional data.
}

