% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{+.fData}
\alias{+.fData}
\title{Overload of + operator for fData objects.}
\usage{
\method{+}{fData}(fD, A)
}
\arguments{
\item{fD}{the univariate functional data object}

\item{A}{either a functional data object, whose time_grid must be the same as
fD's, or a one-dimensional data structure (like 1D-array or raw numeric
vector), or a two-dimensional data structure (like 2D-array or raw numeric
matrix ), that specifies the second term of the sum.
In case of a one-dimensional data structure, the sum is performed element-wise
between each element of the functional dataset fD and A.
In case of a two-dimensional data structure, the sum is performet element-wise
between corresponding elements of fD and A's rows.}
}
\description{
It allows to perform sum between functional data objects or between a
 functional data and other compliant containers, like matrices or vectors,
 representing the pointwise measurements of the second term of the sum.
}

