% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roahd-package.R
\docType{package}
\name{roahd}
\alias{roahd}
\alias{roahd-package}
\title{roahd: RObust Analysis for High dimensional Data.}
\description{
A package meant to collect and provide methods for the analysis of univariate
and multivariate functional datasets through the use of robust methods, with
special focus on computation of depths and outlier detection.
Special care was devoted to the efficient implementation of robust methods,
so that they can be employed also in high-dimensional datasets,
}
\seealso{
Useful links:
\itemize{
  \item \url{https://astamm.github.io/roahd/}
  \item \url{https://github.com/astamm/roahd}
  \item Report bugs at \url{https://github.com/astamm/roahd/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID}) [contributor]

Authors:
\itemize{
  \item Nicholas Tarabelloni \email{nicholas.tarabelloni@polimi.it}
  \item Ana Arribas-Gil \email{aarribas@est-econ.uc3m.es}
  \item Francesca Ieva \email{francesca.ieva@polimi.it}
  \item Anna Maria Paganoni \email{anna.paganoni@polimi.it}
  \item Juan Romo \email{romo@est-econ.uc3m.es}
}

Other contributors:
\itemize{
  \item Francesco Palma \email{frapalma7892@gmail.com} [contributor]
  \item Antonio Elias-Fernandez [contributor]
}

}
\keyword{internal}
