% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{set_alpha}
\alias{set_alpha}
\title{Function to setup alpha value for a set of colors}
\usage{
set_alpha(col, alpha)
}
\arguments{
\item{col}{a vector of colors}

\item{alpha}{the value(s) of alpha for (each of) the colors}
}
\description{
\code{set_alpha} manipulates a vector of color representations in order to
setup the alpha value, and get the desired transparency level
}

