% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{[.fData}
\alias{[.fData}
\title{Overload of [] access operator for univariate functional dataset.}
\usage{
\method{[}{fData}(fD, i, j, as_fData = TRUE)
}
\arguments{
\item{fD}{the univariate functional dataset}

\item{i}{a valid expression to subset rows (observations) of the univariate
functional dataset}

\item{j}{a valid expression to subset columns (measurements) of the univariate
functional dataset}

\item{as_fData}{logical flag to specify if output should be returned as a fData
object containing the request subset, default is TRUE.}
}
\description{
It provides an easy and natural access to subsets of a functional dataset
without having to deal with the inner representation of functional datasets
of fData
}

