% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{*.fData}
\alias{*.fData}
\title{Overload of * operator for fData objects.}
\usage{
\method{*}{fData}(fD, a)
}
\arguments{
\item{fD}{the univariate functional data object}

\item{a}{either a number or a numeric one-dimensional data strcture (array,
matrix or raw vector)}
}
\description{
It allows to perform multiplications between a functional data object and a
 either a numeric variable or numeric one-dimensional data structure.
}

