% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{toListOfValues}
\alias{toListOfValues}
\title{Extracting list of components values from multivariate functional dataset}
\usage{
toListOfValues(mfData)
}
\arguments{
\item{mfData}{the multivariate functional dataset}
}
\description{
It provides an easy way to extract from a multivariate functional dataset
with arbitrary dimensions a list of its time-by-time values stored each into
the corresponding matrix.
}

