% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{toRowMatrixForm}
\alias{toRowMatrixForm}
\title{Conversion to row-matrix form of generic vectors/arrays/matrices}
\usage{
toRowMatrixForm(D)
}
\arguments{
\item{D}{a generic array or atomic vecor to be converted in row-matrix format.}
}
\description{
\code{toRowMatrixForm} manipulates D to obtain a matrix representation such that,
in case of just one observation we have a row vector representation, and in general
we have a consistent output with matrix-structure
}

