% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{area_under_curve}
\alias{area_under_curve}
\title{Area under curve of elements of a univariate functional dataset}
\usage{
area_under_curve(fData)
}
\arguments{
\item{fData}{the functional dataset containing elements whose areas under the
curve have to be computed}
}
\description{
It computes the area under the curve of elements of a univariate functional
 dataset. A plain trapezoidal rule is employed to compute the integral.
}

