% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthgram.R
\name{depthgram}
\alias{depthgram}
\title{Depthgram for univariate and multivariate functional data sets}
\usage{
depthgram(
  Data,
  marginal_outliers = FALSE,
  boxplot_factor = 1.5,
  outliergram_factor = 1.5,
  ids = NULL
)
}
\arguments{
\item{Data}{A \code{\link[base]{list}} of length \code{p} (number of coordinates).
Each element is an \verb{n x N} matrix with \code{n} individuals and \code{N} time points.}

\item{marginal_outliers}{A boolean specifying whether the function should
return shape and amplitude outliers over each dimension. Defaults to
\code{FALSE}.}

\item{boxplot_factor}{A numeric value specifying the inflation factor for
marginal functional boxplots. This is ignored if \code{marginal_outliers == FALSE}. Defaults to \code{1.5}.}

\item{outliergram_factor}{A numeric value specifying the inflation factor for
marginal outliergrams. This is ignored if \code{marginal_outliers == FALSE}.
Defaults to \code{1.5}.}

\item{ids}{A character vector specifying labels for individual observations.
Defaults to \code{NULL}, in which case observations will remain unlabelled.}
}
\value{
An object of class \code{depthgram} which is a list with the following
items:
\itemize{
\item \code{mbd.mei.d}: vector MBD of the MEI dimension-wise.
\item \code{mei.mbd.d}: vector MEI of the MBD dimension-wise.
\item \code{mbd.mei.t}: vector MBD of the MEI time-wise.
\item \code{mei.mbd.t}: vector MEI of the MEI time-wise.
\item \code{mbd.mei.t2}: vector MBD of the MEI time/correlation-wise.
\item \code{mei.mbd.t2}: vector MEI of the MBD time/correlation-wise.
\item \code{shp.out.det}: detected shape outliers by dimension.
\item \code{mag.out.det}: detected magnitude outliers by dimension.
\item \code{mbd.d}: matrix \verb{n x p} of MBD dimension-wise.
\item \code{mei.d}: matrix \verb{n x p} of MEI dimension-wise.
\item \code{mbd.t}: matrix \verb{n x p} of MBD time-wise.
\item \code{mei.t}: matrix \verb{n x p} of MEI time-wise.
\item \code{mbd.t2}: matrix \verb{n x p} of MBD time/correlation-wise
\item \code{mei.t2}: matrix \verb{n x p} of MBD time/correlation-wise.
}
}
\description{
This function computes the three 'DepthGram' representations from a p-variate
functional data set.
}
\examples{
N <- 2e2
P <- 1e3
grid <- seq(0, 1, length.out = P)
Cov <- exp_cov_function(grid, alpha = 0.3, beta = 0.4)

Data <- list()
Data[[1]] <- generate_gauss_fdata(
  N,
  centerline = sin(2 * pi * grid),
  Cov = Cov
)
Data[[2]] <- generate_gauss_fdata(
  N,
  centerline = sin(2 * pi * grid),
  Cov = Cov
)
names <- paste0("id_", 1:nrow(Data[[1]]))

DG <- depthgram(Data, marginal_outliers = TRUE, ids = names)
}
\references{
\itemize{
\item Aleman-Gomez, Y., Arribas-Gil, A., Desco, M. Elias-Fernandez, A., and Romo,
J. (2021). "Depthgram: Visualizing Outliers in High Dimensional Functional
Data with application to Task fMRI data exploration".
}
}
