% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthGram.R
\name{depthGram}
\alias{depthGram}
\title{DepthGram}
\usage{
depthGram(Data, marg.out = FALSE, FB = 1.5, FO = 1.5, ids = NULL)
}
\arguments{
\item{Data}{is a list of length p (nb. of coordinates). Each element is an n*N matrix, n individuals, N time points.}

\item{marg.out}{if TRUE, the function returns shape and magnitude outliers over each dimension.}

\item{FB}{factor for boxplot rule in functional boxplot for marginal magnitude outlier detection (if marg.out==TRUE).}

\item{FO}{factor for boxplot rule in outliergram for marginal magnitude outlier detection (if marg.out==TRUE).}

\item{ids}{labels for individual observations}
}
\value{
A list with the following items:
\itemize{
\item mbd.mei.d - vector MBD of the MEI dimension-wise.
\item mei.mbd.d - vector MEI of the MBD dimension-wise.
\item mbd.mei.t - vector MBD of the MEI time-wise.
\item mei.mbd.t - vector MEI of the MEI time/correlation-wise.
\item mbd.mei.t2 - vector MBD of the MEI time/correlation-wise.
\item mei.mbd.t2 - vector MEI of the MBD time/correlation-wise.
\item shp.out.det - detected shape outliers by dimension.
\item mag.out.det - detected magnitude outliers by dimension.
\item mbd.d - matrix nxp of MBD dimension-wise.
\item mei.d - matrix nxp of MEI dimension-wise.
\item mbd.t - matrix nxp of MBD time-wise.
\item mei.t - matrix nxp of MEI time-wise.
\item mbd.t2 - matrix nxp of MBD time/correlation-wise
\item mei.t2 - matrix nxp of MBD time/correlation-wise.
}
}
\description{
This function computes the 3 DepthGram representations from a p-variate funcional data set.
}
\examples{

N = 2e2
P = 1e3
grid = seq( 0, 1, length.out = P )
Cov = exp_cov_function( grid, alpha = 0.3, beta = 0.4 )

Data <- list()
Data[[1]] <-  generate_gauss_fdata( N, centerline = sin( 2 * pi * grid ), Cov = Cov )
Data[[2]] <- generate_gauss_fdata( N, centerline = sin( 2 * pi * grid ), Cov = Cov )
names <- paste0("id_", 1:nrow(Data[[1]]))

DG <- depthGram(Data, marg.out=TRUE, ids=names)

}
\references{
Aleman-Gomez, Y., Arribas-Gil, A., Desco, M. Elias-Fernandez, A., and Romo, J. (2021). "Depthgram: Visualizing Outliers in High Dimensional Functional Data with application to Task fMRI data exploration". \url{arXiv:2103.08874}
}
