% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_cov_function}
\alias{exp_cov_function}
\title{Exponential covariance function over a grid for univariate functional data}
\usage{
exp_cov_function(time_grid, alpha, beta)
}
\arguments{
\item{time_grid}{a vector of time points}

\item{alpha}{the alpha parameter in the exponential covariance formula}

\item{beta}{the beta parameter in the exponential covariance formula}
}
\description{
\eqn{C( s, t ) = \alpha e^{ - \beta | s - t | }}
}

