% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbplot.R
\name{fbplot}
\alias{fbplot}
\title{Functional boxplot}
\usage{
fbplot(fData, Depths = "MBD", adjust = FALSE, display = TRUE, ...,
  verbose = FALSE)
}
\arguments{
\item{fData}{the univariate functional dataset whose functional boxplot is
desired}

\item{Depths}{either a vector containing the depths for each element of the
dataset, or a string containing the name of the method you want to use to
compute it. In this case the name of the method must be included in the
caller's environment}

\item{adjust}{either FALSE if you would like the default value for the
inflation factor, F = 1.5, to be used, or a list specifying the parameters
required by the adjustment.}

\item{display}{either a logical value indicating wether you want the
functional boxplot to be displayed, or the number of the graphical devices
where you want the boxplot to be plotted.}
}
\description{
\code{fbplot} displays the functional boxplot of a dataset of functional data.
}

