% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_gauss_fdata}
\alias{generate_gauss_fdata}
\title{Functional boxplot}
\usage{
generate_gauss_fdata(M, centerline, Cov = NULL, CholCov = NULL)
}
\arguments{
\item{M}{the number of functional observations to generate}

\item{centerline}{the centerline of the distribution (either mean or median)}

\item{Cov}{covairance operator, in form of a P x P matrix (where P is the number of time points of the discrete grid over which functional data are observed)}

\item{CholCov}{the Cholesky factor of the P x P Covariance matrix}
}
\description{
\code{generate_gauss_data} displays the functional boxplot of a dataset of functional data.
}

