% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BandDepths.R
\name{MBD}
\alias{MBD}
\title{Modified Band Depth of functions in a functional dataset}
\usage{
MBD(Data)
}
\arguments{
\item{Data}{is the dataset of functional observations whose MBD are to be computed. Its rows
should be the functional observations, while the columns must be time-point evaluations of such
observations}
}
\description{
\code{MBD} Computes the Modified Band Depth (MBD) of elements of a functional dataset
}

