% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{mean.fData}
\alias{mean.fData}
\title{\code{mean.fData} method to compute the sample mean of a fData object.}
\usage{
\method{mean}{fData}(fData)
}
\arguments{
\item{fData}{the functional data object representing the dataset}
}
\description{
It computes the \bold{cross-sectional} mean of a univariate functional
dataset, i.e., its time-by-time sample mean.
}

