% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot.fData}
\alias{plot.fData}
\title{Specialised method to plot univariate functional data..}
\usage{
\method{plot}{fData}(x, ...)
}
\arguments{
\item{x}{the univariate functional data object}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
Specialised method to plot univariate functional data..
}

