% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fData.R
\name{plot.mfData}
\alias{plot.mfData}
\title{\code{plot.mfData} specialised method to plot multivariate functional data.}
\usage{
\method{plot}{mfData}(mfData, lty = 1, col = NULL, ylab = NULL,
  main = NULL, ...)
}
\arguments{
\item{mfData}{the multivariate functional data object}

\item{lty}{lty graphical parameter to be used in plotting functions}

\item{col}{colors to be used in plotting functions}

\item{ylab}{either a string or a list of stirngs to be used as y labels for
each window in the plot panel, default is null}

\item{main}{either a string or a list of stirngs to be used as main title for
each window in the plot panel, default is null}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
\code{plot.mfData} specialised method to plot multivariate functional data.
}

