% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roahd.R
\docType{package}
\name{roahd}
\alias{roahd}
\alias{roahd-package}
\title{roahd: RObust Analysis for High dimensional Data.}
\description{
A package meant to collect and provide methods for the analysis of unviariate
and multivariate functional datasets through the use of robust methods, with
special focus on computation of depths and outlier detection.
Special care was devoted to the efficient implementation of robust methods,
so that they can be employed also in high-dimensional datasets,
}
